+++ studio/src/app/providers/sync/sync.ic.service.ts
@@ -1,21 +1,21 @@
 import {Identity} from '@dfinity/agent';
 
-import authStore from '../../../stores/auth.store';
-import syncStore from '../../../stores/sync.store';
-import offlineStore from '../../../stores/offline.store';
+import authStore from '../../stores/auth.store';
+import syncStore from '../../stores/sync.store';
+import offlineStore from '../../stores/offline.store';
 
-import {SyncData} from '../../../types/editor/sync';
-import {InternetIdentityAuth} from '../../../types/core/ic.identity';
+import {SyncData} from '../../types/editor/sync';
+import {InternetIdentityAuth} from '../../types/core/ic.identity';
 
-import {internetComputer} from '../../../utils/core/environment.utils';
-import {internetIdentityAuth} from '../../../utils/core/ic.identity.utils';
+import {internetComputer} from '../../utils/core/environment.utils';
+import {internetIdentityAuth} from '../../utils/core/ic.identity.utils';
 
 import {SyncService} from './sync.service';
 
-import {AuthFactoryProvider} from '../../../providers/auth/auth.factory.provider';
-import {AuthIcProvider} from '../../../providers/auth/auth.ic.provider';
+import {AuthFactoryProvider} from '../auth/auth.factory.provider';
+import {AuthIcProvider} from '../auth/auth.ic.provider';
 
-import {uploadWorker} from '../../../workers/sync.ic.worker';
+import {uploadWorker} from '../../workers/sync.ic.worker';
 
 export class SyncIcService extends SyncService {
   // @Override
